requirements.txt
fastapi
uvicorn
pymunk
requests
motor
python-dotenv
pydantic
